Split console module
Split the console into 3 parts: common (protocol), client and server.
This helps falling into import loops.
